+++ wakamiti-plugins/wakamiti-db/src/main/java/es/iti/wakamiti/database/DatabaseStepContributor.java
@@ -18,6 +18,10 @@ import java.sql.Connection;
 import java.sql.SQLException;
 
 import es.iti.wakamiti.api.WakamitiAPI;
+import es.iti.wakamiti.api.WakamitiRunContext;
+import es.iti.wakamiti.api.WakamitiStepRunContext;
+import es.iti.wakamiti.api.annotations.SetUp;
+import es.iti.wakamiti.api.util.ResourceLoader;
 import es.iti.wakamiti.api.util.WakamitiLogger;
 import org.apache.poi.openxml4j.exceptions.InvalidFormatException;
 import org.hamcrest.Matchers;
@@ -144,9 +148,7 @@ public class DatabaseStepContributor implements StepContributor {
     // TODO: mostrar la ruta absoluta del fichero en el mensaje de error cuando no lo encuentra
     @Step("db.define.cleanup.file")
     public void setManualCleanup(File file) throws IOException {
-        try (Reader reader = new FileReader(file)) {
-            helper.setCleanUpOperations(IOUtils.toString(reader), file.toString());
-        }
+        helper.setCleanUpOperations(readFileAsString(file), file.toString());
     }
 
     @Step(
@@ -172,10 +174,10 @@ public class DatabaseStepContributor implements StepContributor {
 
     @Step("db.action.script.file")
     public void executeSQLScript(File file) throws IOException, SQLException, JSQLParserException {
-        try (Reader reader = new FileReader(file)) {
-            helper.executeSQLStatements(IOUtils.toString(reader), file.toString(), enableCleanupUponCompletion);
-        }
+        helper.executeSQLStatements(readFileAsString(file), file.toString(), enableCleanupUponCompletion);
     }
+
+
 
 
     @Step(value = "db.action.insert.from.data", args = "word")
@@ -192,7 +194,7 @@ public class DatabaseStepContributor implements StepContributor {
 
     @Step("db.action.insert.from.xls")
     public void insertFromXLSFile(File file) throws IOException, SQLException {
-        try (MultiDataSet multiDataSet = new OoxmlDataSet(file, xlsIgnoreSheetRegex, nullSymbol)) {
+        try (MultiDataSet multiDataSet = new OoxmlDataSet(absolutePath(file), xlsIgnoreSheetRegex, nullSymbol)) {
             helper.deleteMultiDataSet(multiDataSet, false);
             helper.insertMultiDataSet(multiDataSet.copy(), enableCleanupUponCompletion);
         }
@@ -201,7 +203,7 @@ public class DatabaseStepContributor implements StepContributor {
 
     @Step(value = "db.action.insert.from.csv", args = { "csv:file", "table:word" })
     public void insertFromCSVFile(File file, String table) throws IOException, SQLException {
-        try (DataSet dataSet = new CsvDataSet(table, file, csvFormat, nullSymbol)) {
+        try (DataSet dataSet = new CsvDataSet(table, absolutePath(file), csvFormat, nullSymbol)) {
             helper.deleteDataSet(dataSet, false);
             helper.insertDataSet(dataSet.copy(), enableCleanupUponCompletion);
         }
@@ -224,7 +226,7 @@ public class DatabaseStepContributor implements StepContributor {
         File file
     ) throws IOException, SQLException, InvalidFormatException {
         try (MultiDataSet multiDataSet = new OoxmlDataSet(
-            file, xlsIgnoreSheetRegex, nullSymbol
+            absolutePath(file), xlsIgnoreSheetRegex, nullSymbol
         )) {
             helper.deleteMultiDataSet(multiDataSet, enableCleanupUponCompletion);
         }
@@ -233,7 +235,7 @@ public class DatabaseStepContributor implements StepContributor {
 
     @Step(value = "db.action.delete.from.csv", args = { "csv:file", "table:word" })
     public void deleteFromCSVFile(File file, String table) throws IOException, SQLException {
-        try (DataSet dataSet = new CsvDataSet(table, file, csvFormat, nullSymbol)) {
+        try (DataSet dataSet = new CsvDataSet(table, absolutePath(file), csvFormat, nullSymbol)) {
             helper.deleteDataSet(dataSet, enableCleanupUponCompletion);
         }
     }
@@ -345,7 +347,7 @@ public class DatabaseStepContributor implements StepContributor {
         File file
     ) throws InvalidFormatException, IOException, SQLException {
         try (MultiDataSet multiDataSet = new OoxmlDataSet(
-            file, xlsIgnoreSheetRegex, nullSymbol
+            absolutePath(file), xlsIgnoreSheetRegex, nullSymbol
         )) {
             helper.assertMultiDataSetExists(multiDataSet);
         }
@@ -354,7 +356,7 @@ public class DatabaseStepContributor implements StepContributor {
 
     @Step(value = "db.assert.table.exists.csv", args = { "csv:file", "table:word" })
     public void assertCSVFileExists(File file, String table) throws IOException, SQLException {
-        try (DataSet dataSet = new CsvDataSet(table, file, csvFormat, nullSymbol)) {
+        try (DataSet dataSet = new CsvDataSet(table, absolutePath(file), csvFormat, nullSymbol)) {
             helper.assertDataSetExists(dataSet);
         }
     }
@@ -486,7 +488,7 @@ public class DatabaseStepContributor implements StepContributor {
         File file
     ) throws InvalidFormatException, IOException, SQLException {
         try (MultiDataSet multiDataSet = new OoxmlDataSet(
-            file, xlsIgnoreSheetRegex, nullSymbol
+            absolutePath(file), xlsIgnoreSheetRegex, nullSymbol
         )) {
             helper.assertMultiDataSetNotExists(multiDataSet);
         }
@@ -495,10 +497,12 @@ public class DatabaseStepContributor implements StepContributor {
 
     @Step(value = "db.assert.table.not.exists.csv", args = { "csv:file", "table:word" })
     public void assertCSVFileNotExists(File file, String table) throws IOException, SQLException {
-        try (DataSet dataSet = new CsvDataSet(table, file, csvFormat, nullSymbol)) {
+        try (DataSet dataSet = new CsvDataSet(table, absolutePath(file), csvFormat, nullSymbol)) {
             helper.assertDataSetNotExists(dataSet);
         }
     }
+
+
 
 
     @Step(value = "db.assert.table.empty", args = "word")
@@ -513,4 +517,17 @@ public class DatabaseStepContributor implements StepContributor {
     }
 
 
+    private File absolutePath(File file) {
+        return resourceLoader().absolutePath(file);
+    }
+
+
+    private String readFileAsString(File file) {
+        return resourceLoader().readFileAsString(file);
+    }
+
+
+    private static ResourceLoader resourceLoader() {
+        return WakamitiRunContext.current().resourceLoader();
+    }
 }