+++ wakamiti/wakamiti-core/src/main/java/es/iti/wakamiti/core/Wakamiti.java
@@ -47,7 +47,7 @@ public class Wakamiti {
 
     public static final Logger LOGGER = WakamitiLogger.forClass(Wakamiti.class);
 
-    private static final ResourceLoader resourceLoader = new ResourceLoader();
+    private static final ResourceLoader resourceLoader = new ResourceLoader(new File("."));
     private static final WakamitiContributors contributors = new WakamitiContributors();
     private static final PlanSerializer planSerializer = new JsonPlanSerializer();
     private static final EventDispatcher eventDispatcher = new EventDispatcher();
@@ -66,8 +66,19 @@ public class Wakamiti {
     }
 
 
+    @Deprecated
     public static ResourceLoader resourceLoader() {
         return resourceLoader;
+    }
+
+
+    public static ResourceLoader resourceLoader(File workingDir) {
+        return new ResourceLoader(workingDir);
+    }
+
+
+    public static ResourceLoader resourceLoader(Configuration configuration) {
+        return new ResourceLoader(workingDir(configuration));
     }
 
 
@@ -103,6 +114,12 @@ public class Wakamiti {
     /** @return The default configuration. Any configuration should be derived from this one */
     public static Configuration defaultConfiguration() {
         return WakamitiConfiguration.DEFAULTS;
+    }
+
+
+    /** @return The working directory for a given configuration*/
+    public static Path workingDir(Configuration configuration) {
+        return Path.of(configuration.get(WORKING_DIR,String.class).orElse("")).toAbsolutePath();
     }
 
 
@@ -352,11 +369,13 @@ public class Wakamiti {
 
 
     public PlanNode executePlan(PlanNode plan, Configuration configuration) {
+        WakamitiRunContext.set(new WakamitiRunContext(configuration));
         PlanNode result = new PlanRunner(plan, configuration).run();
         writeOutputFile(plan, configuration);
         if (configuration.get(WakamitiConfiguration.REPORT_GENERATION, Boolean.class).orElse(true)) {
             generateReports(configuration,new PlanNodeSnapshot(plan));
         }
+        WakamitiRunContext.clear();
         return result;
     }
 
@@ -407,7 +426,9 @@ public class Wakamiti {
 
     private Path writeStandardOutputFile(PlanNode plan, Configuration configuration) throws IOException {
         String outputPath = configuration.get(WakamitiConfiguration.OUTPUT_FILE_PATH, String.class).orElseThrow();
-        Path path = PathUtil.replacePlaceholders(Paths.get(outputPath).toAbsolutePath(), plan);
+        Path path = resourceLoader(configuration).absolutePath(
+            PathUtil.replacePlaceholders(Paths.get(outputPath), plan)
+        );
         Path parentPath = path.getParent();
         if (parentPath != null) {
             Files.createDirectories(parentPath);
@@ -424,7 +445,9 @@ public class Wakamiti {
 
     private void writeOutputFilesPerTestCase(PlanNode plan, Configuration configuration) throws IOException {
         String outputPath = configuration.get(OUTPUT_FILE_PER_TEST_CASE_PATH, String.class).orElseThrow();
-        Path path = PathUtil.replacePlaceholders(Paths.get(outputPath).toAbsolutePath(), plan);
+        Path path = resourceLoader(configuration).absolutePath(
+            PathUtil.replacePlaceholders(Paths.get(outputPath), plan)
+        );
         Files.createDirectories(path);
 
         List<PlanNode> testCases = plan
@@ -452,7 +475,7 @@ public class Wakamiti {
                 REPORT_SOURCE
             );
         }
-        generateReports(configuration, Path.of(reportSource));
+        generateReports(configuration, resourceLoader(configuration).absolutePath(Path.of(reportSource)));
     }