+++ wakamiti-plugins/wakamiti-cucumber-exporter/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [2.1.0] - 2023-07-05
+
+### Modified
+
+- Use absolute file paths based on working dir
+
+
 ## [2.0.0] - 2023-05-22
 
 ### Modified